Move `save(task)` into a `CreateTaskUseCase` component
